Added automatic wall detection
Each node now detects if a wall/obstacle is in its space. If so, the node will be marked as wall and this node will not be available for the algorithm to go to.

+++ Assets/Scripts/Pathfinding/AlgorithmTester.cs
@@ -6,7 +6,7 @@ class AlgorithmTester : MonoBehaviour
 {
     [SerializeField] int _gridWidth = 5;
     [SerializeField] int _gridHeight = 5;
-    [SerializeField] int _gridNodeSize = 1;
+    [SerializeField] float _gridNodeSize = 1;
 
     [SerializeField] Transform _testStartPosition;
     [SerializeField] Transform _testEndPosition;
